Tutorial Screen updates

+++ app/screens/RoomScreen.js
@@ -26,7 +26,6 @@ import { MenuButton } from '../components/MenuButton.js';
 
 import { isInGame } from "../auth";
 import { isInRoom } from "../auth";
-import FadeInView from '../components/FadeInView.js';
 import colors from '../misc/colors.js';
 import Rolesheet from '../misc/roles.json';
 
@@ -52,19 +51,12 @@ class Room_Screen extends React.Component {
         firebase.database().ref('rooms/' + roomname).set({
             phase: 1,
             owner: firebase.auth().currentUser.uid,
-            playernum: 1,
             daycounter:1,
         });
-
-        firebase.database().ref('owners').child(firebase.auth().currentUser.uid)
-            .update({
-                roomname: roomname,
-            })
 
         //Set up list of players
         firebase.database().ref('rooms/' + roomname + '/listofplayers/' 
             + firebase.auth().currentUser.uid).set({
-                name:               'Owner',
                 actionbtnvalue:     false,
                 presseduid:         'foo',
         });
@@ -105,115 +97,6 @@ class Room_Screen extends React.Component {
             />
             <View style = {{flex:0.02}}/>
         </View>
-    }
-}
-
-class Create_Screen extends React.Component {
-
-    constructor(props) {
-        super(props);
-
-        this.state = {
-            alias:'',
-        };
-    }
-
-    _createRoom() {
-        const roomname = randomize('A',4);
-
-        //TODO: Check if room already exists
-        
-        AsyncStorage.setItem('ROOM-KEY', roomname);
-        
-        firebase.database().ref('rooms/' + roomname).set({
-            phase: 1,
-            owner: firebase.auth().currentUser.uid,
-            playernum: 1,
-            daycounter:1,
-        });
-
-        //Set up list of players
-        firebase.database().ref('rooms/' + roomname + '/listofplayers/' 
-            + firebase.auth().currentUser.uid).set({
-                name:               this.state.alias,
-                actionbtnvalue:     false,
-                presseduid:         'foo',
-        });
-        
-        firebase.database().ref('listofroles/'+firebase.auth().currentUser.uid).update({b:0})
-
-        this.props.navigation.dispatch(
-            NavigationActions.reset({
-                index: 0,
-                actions: [
-                  NavigationActions.navigate({ routeName: 'Lobby_Screen', params:{roomname:roomname}})
-                ]
-            })
-        )
-        Keyboard.dismiss();
-    }
-
-    _validName(name) {
-        if(name.length>0 && name.length < 11){
-            return true
-        } else {
-            alert('Name is not a valid length (10 characters or less)')
-            return false
-        }
-    }
-
-    render() {
-        return <TouchableWithoutFeedback 
-        style = {{ flex:1 }}
-        onPress={()=>{ Keyboard.dismiss() }}>
-            <View style = {{flex:1,backgroundColor:colors.background}}>
-                <View style = {{flex:0.5,backgroundColor:colors.background, 
-                    justifyContent:'center', alignItems:'center'}}>
-
-                    <View style ={{flex:0.4}}/>
-
-                    <TouchableOpacity
-                        style = {{flexDirection:'row'}}
-                        onPress = {()=>{
-                            this.props.navigation.goBack();
-                        }} >
-                        <MaterialCommunityIcons name='close'
-                            style={{color:colors.font,fontSize:40}}/>
-                    </TouchableOpacity>
-
-                    <View style = {{ flexDirection: 'row'}}>
-                        <TextInput
-                            placeholder="Who are you?"
-                            placeholderTextColor={colors.main}
-                            style={{
-                                backgroundColor: colors.background,
-                                flex:0.6,
-                                fontFamily:'ConcertOne-Regular',
-                                fontSize: 20,
-                                color:colors.font,
-                                textAlign:'center',
-                            }}
-                            value={this.state.alias}
-                            onChangeText = {(text) => {this.setState({alias: text})}}
-                            focus={true}
-                            onSubmitEditing = {()=>{Keyboard.dismiss()}}
-                        />
-                    </View>
-
-                    <MenuButton
-                        title="Create Room"
-                        flex = {0.75}
-                        fontSize={20}
-                        onPress={()=>{
-                            if(this._validName(this.state.alias)){
-                                this._createRoom()
-                            }
-                        }}
-                    />
-
-                </View>
-            </View>
-        </TouchableWithoutFeedback>
     }
 }
 
@@ -365,7 +248,6 @@ class Lobby_Screen extends React.Component {
 
         this.state = {
             roomname:       params.roomname.toUpperCase(),
-            xdisabled:      true,
             namelist:       [],
             rolecount:      0,
             playercount:    0,
@@ -383,39 +265,39 @@ class Lobby_Screen extends React.Component {
     componentWillMount() {
 
         this.roomRef.on('value',roomsnap=>{
-
-            this.listRef.once('value',snap => {
-                var list = [];
-                snap.forEach((child)=> {
-                    list.push({
-                        name: child.val().name,
-                        owner: roomsnap.val().owner == child.key,
-                        key: child.key,
+            if(roomsnap.exists()){
+                this.listRef.once('value',snap => {
+                    var list = [];
+                    snap.forEach((child)=> {
+                        list.push({
+                            name: child.val().name,
+                            owner: roomsnap.val().owner == child.key,
+                            key: child.key,
+                        })
                     })
+                    this.setState({namelist:list,playercount:snap.numChildren()})
                 })
-                this.setState({namelist:list,playercount:snap.numChildren()})
-            })
 
-            if(roomsnap.val().phase > 1 ){
-                AsyncStorage.setItem('GAME-KEY',this.state.roomname);
+                if(roomsnap.val().phase > 1 ){
+                    AsyncStorage.setItem('GAME-KEY',this.state.roomname);
 
-                this.props.navigation.dispatch(
-                    NavigationActions.reset({
-                        index: 0,
-                        actions: [
-                            NavigationActions.navigate({ routeName: 'Mafia_Screen',
-                            params:{roomname:this.state.roomname}})
-                        ]
-                    })
-                )
+                    this.props.navigation.dispatch(
+                        NavigationActions.reset({
+                            index: 0,
+                            actions: [
+                                NavigationActions.navigate({ routeName: 'Mafia_Screen',
+                                params:{roomname:this.state.roomname}})
+                            ]
+                        })
+                    )
+                }
+
+                if(roomsnap.val().owner == firebase.auth().currentUser.uid){
+                    this.setState({amiowner:true})
+                } else {
+                    this.setState({amiowner:false})
+                }
             }
-
-            if(roomsnap.val().owner == firebase.auth().currentUser.uid){
-                this.setState({amiowner:true})
-            } else {
-                this.setState({amiowner:false})
-            }
-
         })
 
         
@@ -811,9 +693,6 @@ export const createRoomNavigator = (inGame,inRoom,isExpired,key) => {
         {
             Room_Screen: {
                 screen: Room_Screen,
-            },
-            Create_Screen: {
-                screen: Create_Screen,
             },
             Join_Screen: {
                 screen: Join_Screen,
